requirements.txt
fastapi
uvicorn
streamlit
pydantic
pydantic-settings
email-validator
requests
sqlalchemy
psycopg2-binary
pgvector
python-dotenv
openai
fpdf2
plotly
pandas
openpyxl
passlib[bcrypt]
google-search-results
pypdf
langchain-text-splitters
python-multipart